Added relationship statements
Added relationship statements to each object where appropriate.
Added user_id property to votes in order to be able to tell which user made which vote.
I've made the loading lazy for the user object (since when you query for a user you won't always need to load their posts, comments, etc.) but threads and comments will load eagerly (since you probably want to be able to see the vote count on each). 

I think I got it right but there's a decent chance I messed something up but it's 2am and that's my only excuse.